+++ classes/course_renderer.php
@@ -265,9 +265,14 @@ class format_topcoll_course_renderer extends \core_course_renderer {
             return '';
         }
 
+        // If module is not visible to the user then don't bother getting meta data.
+        if (!$mod->uservisible) {
+            return '';
+        }
+
         // Do we have an activity function for this module for returning meta data?
         $meta = \format_topcoll\activity::module_meta($mod);
-        if (!$meta->is_set(true)) {
+        if (($meta == null) || (!$meta->is_set(true))) {
             // Can't get meta data for this module.
             return '';
         }
@@ -354,11 +359,11 @@ class format_topcoll_course_renderer extends \core_course_renderer {
             $engagementmeta = array();
 
             // Below, !== false means we get 0 out of x submissions.
-            if (!$meta->submissionnotrequired && $meta->numsubmissions !== false) {
+            if (!$meta->submissionnotrequired && $meta->numparticipants !== false) {
                 $engagementmeta[] = get_string('xofy'.$meta->submitstrkey, 'format_topcoll',
                     (object) array(
                         'completed' => $meta->numsubmissions,
-                        'participants' => \format_topcoll\toolbox::course_participant_count($COURSE->id, $mod->modname)
+                        'participants' => $meta->numparticipants
                     )
                 );
             }