+++ Changes.md
@@ -5,6 +5,7 @@ Version 3.8.1.2 - TBD
   2. Fix logic in restore of backup file for 'numsections' - #73.
   3. Colour fixes for Adaptable theme version 2.3.1.
   4. Fix 'Show hidden sections in collapsed form does not reveal section titles to students' - CONTRIB-7510.
+  5. Port of fixes to activity meta from Adaptable - https://gitlab.com/jezhops/moodle-theme_adaptable/-/issues/161.
 
 Version 3.8.1.1 - 29/06/2020
   1. Added 'Course page further information' functionality from Adaptable theme.